w32: Fix cast to internal fd.
* src/estream.c (tmpfd): Use intptr_t.

+++ src/estream.c
@@ -4564,7 +4564,7 @@ tmpfd (void)
 #ifdef HAVE_W32CE_SYSTEM
           int fd = (int)file;
 #else
-          int fd = _open_osfhandle ((long)file, 0);
+          int fd = _open_osfhandle ((intptr_t)file, 0);
           if (fd == -1)
             {
               CloseHandle (file);
